fixed bug4

+++ database/migrations/2021_08_02_155350_create_invoices_table.php
@@ -26,7 +26,7 @@ class CreateInvoicesTable extends Migration
             $table->string('postal_code')->nullable()->comment('کد پستی');
             $table->string('national_code')->nullable()->comment('شناسه ملی');
             $table->string('registration_number')->nullable()->comment('شماره ثبت');
-            $table->integer('tax')->nullable()->comment('مالیاات بر ارزش افزوده');
+            $table->bigInteger('tax')->nullable()->comment('مالیاات بر ارزش افزوده');
 
             $table->bigInteger('total')->nullable()->comment(' جمع به ریال');
             $table->bigInteger('final_total')->nullable()->comment('مبلغ کل پس از اعمال ارزش افزوده به ریال');
